+++ CHANGELOG.md
@@ -7,6 +7,13 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [unreleased]
  ### Added
+ ### Changed
+- Improved template cache configuration [#1618](https://github.com/ehrbase/ehrbase/pull/1618)
+- Improved performance of template service [#1618](https://github.com/ehrbase/ehrbase/pull/1618)
+ ### Fixed 
+
+## [2.30.1]
+ ### Added
  ### Changed 
  ### Fixed 
 
@@ -286,4 +293,5 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 [2.28.0]: https://github.com/ehrbase/ehrbase/compare/v2.27.0...v2.28.0
 [2.29.0]: https://github.com/ehrbase/ehrbase/compare/v2.28.0...v2.29.0
 [2.30.0]: https://github.com/ehrbase/ehrbase/compare/v2.29.0...v2.30.0
-[unreleased]: https://github.com/ehrbase/ehrbase/compare/v2.30.0...HEAD
+[2.30.1]: https://github.com/ehrbase/ehrbase/compare/v2.30.0...v2.30.1
+[unreleased]: https://github.com/ehrbase/ehrbase/compare/v2.30.1...HEAD